move `compute complexity` to its own directory.
