+++ cmd/registry/cmd/compute/complexity/complexity-proto.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package compute
+package complexity
 
 import (
 	"archive/zip"
